requirements.txt
selenium
argparse
Pillow
httpx
httpx-socks
python-dotenv
